+++ remoulade/errors.py
@@ -86,3 +86,8 @@ class NoStateBackend(BrokerError):
 class ChannelPoolTimeout(BrokerError):
     """Raised when the broker has wait for tool long to fetch a channel from its channel pool.
     """
+
+
+class InvalidProgress(RemouladeError):
+    """Raised when trying to set a progress that is greater than 1 or less than 0
+    """